requirements.txt
gunicorn==20.0.4
black
pre-commit
requests
pymongo
flask_pymongo
stravalib
python-dotenv==0.10.3
flask_cors
flask